+++ src/common/NoDataMessage.vue
@@ -24,14 +24,20 @@ withDefaults(
 
 <style lang="scss" scoped>
 .no-data-message {
-  display: grid;
-  place-items: center;
-  grid-gap: toRem(12);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: toRem(12);
 }
 
 .no-data-message__icon {
   color: var(--text-primary-light);
   font-size: toRem(48);
+
+  @include respond-to(small) {
+    width: toRem(120);
+    height: toRem(120);
+  }
 }
 
 .no-data-message__message {