+++ src/composables/providers/use-metamask.ts
@@ -8,7 +8,6 @@ import {
   handleEthError,
   requestAddEthChain,
   requestSwitchEthChain,
-  isMobile,
 } from '@/helpers'
 import { computed, ref } from 'vue'
 import {
@@ -24,12 +23,10 @@ import {
 import { Deferrable } from '@ethersproject/properties'
 import { TransactionRequest } from '@ethersproject/abstract-provider'
 import { useNetworksStore } from '@/store'
-import { router } from '@/router'
 
 export const useMetamask = (provider: ProviderInstance): ProviderWrapper => {
   const chainId = ref<ChainId>('')
   const selectedAddress = ref('')
-  const METAMASK_APP_CONNECT_URL = `https://metamask.app.link/dapp/${window.location.host}${router.currentRoute.value.fullPath}`
 
   const currentProvider = computed(
     () =>
@@ -79,11 +76,6 @@ export const useMetamask = (provider: ProviderInstance): ProviderWrapper => {
 
   const connect = async () => {
     try {
-      if (isMobile() && !navigator.userAgent.includes('MetaMask')) {
-        window.open(METAMASK_APP_CONNECT_URL)
-        return
-      }
-
       await connectEthAccounts(currentProvider.value)
     } catch (error) {
       handleEthError(error as EthProviderRpcError)